Revert "ACMS-4504: Remove Acquia CMS Common Dependency from Drupal Starter Kit Headless module. (#1930)"
This reverts commit d8ee4a3d300e6d85363fa4f23a577265e3d62131.
