+++ modules/acquia_cms_headless/composer.json
@@ -4,8 +4,8 @@
     "license": "GPL-2.0-or-later",
     "type": "drupal-module",
     "require": {
-        "drupal/acquia_cms_tour": "2.x-dev",
-        "drupal/entity_clone": "^2.0",
+        "drupal/acquia_cms_common": "^1.9 || ^2.1 || ^3.1",
+        "drupal/acquia_cms_tour": "^2.1.8",
         "drupal/jsonapi_extras": "^3.26",
         "drupal/jsonapi_menu_items": "^1.2",
         "drupal/next": "^2.0",
@@ -15,8 +15,7 @@
         "drupal/restui": "^1.21"
     },
     "require-dev": {
-        "drupal/acquia_claro": "^1.3",
-        "drupal/acquia_cms_toolbar": "1.x-dev"
+        "drupal/acquia_claro": "^1.3"
     },
     "conflict": {
         "drupal/consumers": "<1.16",
