+++ modules/acquia_cms_headless/config/install/user.role.headless.yml
@@ -18,3 +18,5 @@ permissions:
   - 'bypass node access'
   - 'issue subrequests'
   - 'view all revisions'
+  - 'view any unpublished content'
+  - 'view latest version'
