--- modules/acquia_cms_headless/config/optional/user.role.administrator.yml
@@ -1,8 +1,0 @@
-langcode: en
-status: true
-dependencies: {  }
-id: administrator
-label: Administrator
-weight: -10
-is_admin: true
-permissions: {  }
