+++ modules/acquia_cms_headless/modules/acquia_cms_headless_ui/acquia_cms_headless_ui.info.yml
@@ -4,6 +4,5 @@ type: module
 core_version_requirement: ^10 || ^11
 description: 'Implements Pure Headless mode, disabling the Drupal Front-End and updating the admin experience.'
 dependencies:
-  - acquia_cms_headless:acquia_cms_headless
-  - drupal:content_moderation
+  - drupal:acquia_cms_headless
   - drupal:path_alias
