+++ modules/acquia_cms_headless/tests/src/Functional/DashboardApiUsersTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\acquia_cms_headless\FunctionalJavascript;
+namespace Drupal\Tests\acquia_cms_headless\Functional;
 
 use Drupal\Tests\acquia_cms_headless\Traits\DashboardSectionTrait;
 use Drupal\Tests\acquia_cms_headless\Traits\DashboardTableTrait;
@@ -52,30 +52,30 @@ class DashboardApiUsersTest extends HeadlessTestBase {
     /** @var \Drupal\FunctionalJavascriptTests\JSWebAssert $assertSession */
     $assertSession = $this->assertSession();
 
-    // Test API Users section exists, get an API Users section.
+    // Test API Users section exists, get API Users section.
     $usersFieldset = $assertSession->waitForElementVisible('css', $this->sectionSelector);
 
     // Test add API user button link has destination.
     $this->assertButtonLink($usersFieldset, '/admin/people/create?destination=/admin/headless/dashboard');
 
-    // Test table body exists and has data in the same order.
-    $this->assertEquals('Headless', $this->getTableBodyColumn(0, 1)->getText());
+    // Test table body exist and has data in same order.
+    $this->assertEquals('Headless', $this->getTableBodyColumn(0)->getText());
     $this->assertEquals('Headless Administrator', $this->getTableBodyColumn(1)->getText());
 
     // Get the API Users operations dropdown elements.
     $dropdownList = $usersFieldset->findAll('css', 'tbody tr:nth-child(1) ul li a');
     $this->assertCount(3, $dropdownList);
 
-    // Click on the Edit button.
+    // Click on Edit button.
     $expectedUrl = $this->baseUrl . '/user/2/edit?destination=/admin/headless/dashboard';
     $usersFieldset->findButton('List additional actions')->click();
-    $this->testOperation($usersFieldset, 'Edit', $expectedUrl, 'Access denied');
+    $this->testOperation($usersFieldset, 'Edit', $expectedUrl, 'Whoops!');
 
     // Click on Clone button.
     $this->drupalGet("/admin/headless/dashboard");
     $usersFieldset->findButton('List additional actions')->click();
     $expectedUrl = $this->baseUrl . '/entity_clone/user/2?destination=/admin/headless/dashboard';
-    $this->testOperation($usersFieldset, 'Clone', $expectedUrl, 'Access denied');
+    $this->testOperation($usersFieldset, 'Clone', $expectedUrl, 'Whoops!');
   }
 
   /**
