+++ modules/acquia_cms_headless/tests/src/Functional/DashboardNextjsSitesTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\acquia_cms_headless\FunctionalJavascript;
+namespace Drupal\Tests\acquia_cms_headless\Functional;
 
 use Drupal\Tests\acquia_cms_headless\Traits\DashboardSectionTrait;
 use Drupal\Tests\acquia_cms_headless\Traits\DashboardTableTrait;
@@ -52,16 +52,16 @@ class DashboardNextjsSitesTest extends HeadlessTestBase {
     /** @var \Drupal\FunctionalJavascriptTests\JSWebAssert $assertSession */
     $assertSession = $this->assertSession();
     $assertSession->waitForElementVisible('css', $this->sectionSelector);
-    // Test API Keys section exists, get an API Keys section.
+    // Test API Keys section exists, get API Keys section.
     $nextjsSitesFieldset = $this->getSection();
 
-    // Test create a new consumer button link has destination.
+    // Test create new consumer button link has destination.
     $this->assertButtonLink($nextjsSitesFieldset, '/admin/config/services/next/sites/add?destination=/admin/headless/dashboard');
 
     // Click on Add Next.js site button.
     $this->testAddNextjsSite($nextjsSitesFieldset);
 
-    // Test table body exists and has data in the same order.
+    // Test table body exist and has data in same order.
     $this->drupalGet("/admin/headless/dashboard");
     $this->assertEquals('No next.js sites currently exist.', $this->getTableBodyColumn(0)->getText());
 
@@ -78,16 +78,16 @@ class DashboardNextjsSitesTest extends HeadlessTestBase {
     /** @var \Drupal\FunctionalJavascriptTests\JSWebAssert $assertSession */
     $assertSession = $this->assertSession();
     $assertSession->waitForElementVisible('css', $this->sectionSelector);
-    // Test API Keys section exists, get an API Keys section.
+    // Test API Keys section exists, get API Keys section.
     $nextjsSitesFieldset = $this->getSection();
 
-    // Test create a new consumer button link has destination.
+    // Test create new consumer button link has destination.
     $this->assertButtonLink($nextjsSitesFieldset, '/admin/config/services/next/sites/add?destination=/admin/headless/dashboard');
 
     // Create Next.js Site using administrator user.
     $this->testAddNextjsSiteAdmin($nextjsSitesFieldset);
 
-    // Test table body exists and has data in the same order.
+    // Test table body exist and has data in same order.
     $this->assertEquals('headless', $this->getTableBodyColumn(0)->getText());
     $this->assertEquals('headless', $this->getTableBodyColumn(1)->getText());
     $this->assertEquals('http://localhost:3000', $this->getTableBodyColumn(2)->getText());
@@ -96,26 +96,26 @@ class DashboardNextjsSitesTest extends HeadlessTestBase {
     $dropdownList = $nextjsSitesFieldset->findAll('css', 'ul li a');
     $this->assertCount(5, $dropdownList);
 
-    // Click on the Environment variables button.
+    // Click on Environment variables button.
     $this->testEnvironmentVariables($nextjsSitesFieldset);
 
     // Click on Clone button.
     $this->testClone($nextjsSitesFieldset);
 
-    // Test cloned next.js site exists.
+    // Test cloned next.js site exist.
     $this->assertCount(2, $nextjsSitesFieldset->findAll('css', 'table tbody tr'));
     $this->assertEquals('headless clone', $nextjsSitesFieldset->find('css', 'tbody > tr.even > td:nth-child(2)')->getText());
     $nextjsSitesFieldsetCloned = $nextjsSitesFieldset->find('css', 'tbody > tr.even');
 
-    // Click the Edit button.
+    // Click Edit button.
     $this->testEdit($nextjsSitesFieldsetCloned);
     $this->assertEquals('headless clone edit', $nextjsSitesFieldset->find('css', 'tbody > tr.even > td:nth-child(2)')->getText());
 
-    // Click on the Delete button.
+    // Click on Delete button.
     $this->testDelete($nextjsSitesFieldsetCloned);
     $this->assertCount(1, $nextjsSitesFieldset->findAll('css', 'table tbody tr'));
 
-    // Click on the New preview secrete button.
+    // Click on New preview secrete button.
     $this->testNewPreviewSecret($nextjsSitesFieldset);
 
   }
@@ -127,7 +127,7 @@ class DashboardNextjsSitesTest extends HeadlessTestBase {
     $this->assertSession()->elementExists('named', ['link', 'Add Next.js site'], $nextjsSitesFieldset)->click();
     $page = $this->getSession()->getPage();
     $this->assertNotEmpty($page);
-    $this->assertSession()->pageTextContains('Access denied');
+    $this->assertSession()->pageTextContains('Access denied!');
     $expectedUrl = $this->baseUrl . '/admin/config/services/next/sites/add?destination=/admin/headless/dashboard';
     $this->assertSame($expectedUrl, $this->getSession()->getCurrentUrl());
   }
