+++ modules/acquia_cms_headless/tests/src/Functional/HeadlessDrushCommandsTest.php
@@ -30,6 +30,21 @@ class HeadlessDrushCommandsTest extends BrowserTestBase {
   protected $defaultTheme = 'stark';
 
   /**
+   * Disable strict config schema checks in this test.
+   *
+   * Scheduler has a config schema errors, and until it's fixed,
+   * this test cannot pass unless we disable strict config schema checking
+   * altogether. Since strict config schema isn't critically important in
+   * testing this functionality, it's okay to disable it for now, but it should
+   * be re-enabled (i.e., this property should be removed) as soon as possible.
+   *
+   * @var bool
+   */
+  // @codingStandardsIgnoreStart
+  protected $strictConfigSchema = FALSE;
+  // @codingStandardsIgnoreEnd
+
+  /**
    * {@inheritdoc}
    */
   public function testHeadlessCommands(): void {
